+++ src/public/js/app/endpoints-view.js
@@ -1,27 +1,29 @@
 /**
  * Created by jpsoroulas.
  */
-define(['underscore', 'jquery', 'backbone', 'endpoint-view', 'templates', 'jquery-ui'], function (_, $, Backbone, EndpointView, JST) {
+define(['underscore', 'jquery', 'backbone', 'endpoint-view', 'endpoint-history', 'templates', 'jquery-ui'], function (_, $, Backbone, EndpointView, EndpointHistory, JST) {
 
     return Backbone.View.extend({
         el: '#endpoints-table',
         //tagName: 'div',
-        template: JST['src/views/endpoints.hbs'],
+        template: JST['src/views/endpoints-table.hbs'],
 
         initialize: function () {
             this.listenTo(this.collection, 'reset', this.render);
             this.listenTo(this.collection, 'remove', this.render);
             this.listenTo(this.collection, 'add', this.render);
+            this.endpointHistory = new EndpointHistory();
         },
 
         render: function () {
+            var _self = this;
             this.$el.empty();
             var endpoints = this.collection.models;
             var html = this.template(this._computeStats(endpoints));
             this.$el.html(html);
             var subView = document.createDocumentFragment();
             _.each(endpoints, function (endpoint) {
-                var endpointView = new EndpointView({model: endpoint});
+                var endpointView = new EndpointView({model: endpoint, endpointHistory: _self.endpointHistory});
                 subView.appendChild(endpointView.render().el);
             }, this);
             this.$el.find('tbody').append(subView);