+++ src/watcher.js
@@ -55,6 +55,8 @@
  *  }
  *  ```
  *  * [__exportDir__] (_<`project directory`>/storage/_), the storage directory of dynamically created endpoints.
+ *  * [__dbConnectionURL__] (_'mongodb://localhost:27017/'_), the connection url of history persistent storage
+ *  (currently, the mongoDB is used)
  *
  *
  * * Endpoint
@@ -294,6 +296,8 @@ var events = require('events');
 var path = require('path');
 var stampit = require('stampit');
 var _ = require('underscore');
+var async = require('async');
+var asyncTimeout = require('async-timeout');
 var moment = require('moment');
 var nodemailer = require('nodemailer');
 var directTransport = require('nodemailer-direct-transport');
@@ -301,7 +305,6 @@ var directTransport = require('nodemailer-direct-transport');
 //var query = require('connect-query');
 //var fs = require('fs');
 //var connect = require('connect');
-//var async = require('async');
 //var s = require('underscore.string');
 
 var logger = require('./logger');
@@ -309,6 +312,7 @@ var validator = require('./validator');
 var constants = require('./constants');
 var connectors = require('./connectors');
 var httpServerFactory = require('./http-server');
+var dbConnFactory = require('./database');
 //var utils = require('./utils');
 
 var up = constants.serviceStatus.up;
@@ -451,9 +455,12 @@ watcher = stampit().state({
 
         _onStatusResolve: function _onStatusResolve(endpoint, status) {
             if (!this._stopped) {
-                endpoint.timestamp = moment.utc();
+                var now = moment.utc();
+                endpoint.timestamp = now;
                 if (status !== endpoint.status) {
-                    endpoint.since = endpoint.timestamp;
+                    endpoint.since = now;
+                    //store to database
+                    this._storeStatusToHistory(endpoint, status);
                 }
                 endpoint.previousStatus = endpoint.status;
                 endpoint.status = status;
@@ -467,6 +474,22 @@ watcher = stampit().state({
                     setTimeout(_.bind(this._pollEndpoints, this), this.options.interval).unref();
                 }
             }
+        },
+
+        _storeStatusToHistory: function (endpoint, status) {
+            var now = moment.utc();
+            this.db.history.insert({
+                endpointId: endpoint.id,
+                timestamp: now.valueOf(),
+                statusTransition: {
+                    from: endpoint.status,
+                    to: status
+                }
+            }, function (err, docs) {
+                if (err) {
+                    logger.warn('Unable to update the persistent state of \'' + endpoint.id + '\', ' + err);
+                }
+            });
         },
 
         _pollEndpoints: function _pollEndpoints() {
@@ -519,7 +542,7 @@ watcher = stampit().state({
 
         /**
          * Stores endpoint at the persistence storage
-         * (at the moment, the filesystem is used).
+         * (the filesystem is used for easy modification).
          *
          * @private
          * @method _persistEndpointConfig
@@ -539,7 +562,7 @@ watcher = stampit().state({
 
         /**
          * Deletes the endpoint of the specified id from the persistence storage
-         * (at the moment, the filesystem is used).
+         * (the filesystem is used for easy modification).
          *
          * @private
          * @method _deleteEndpoint
@@ -565,7 +588,7 @@ watcher = stampit().state({
 
         /**
          * Updates the endpoint of the specified id from the persistence storage
-         * (at the moment, the filesystem is used).
+         * (the filesystem is used for easy modification).
          *
          * @private
          * @method _updateEndpointConfig
@@ -595,7 +618,7 @@ watcher = stampit().state({
         },
 
         /**
-         * Stores endpoint info at persistence storage (currently, at the filesystem).
+         * Stores endpoint info at persistence storage (at the filesystem for easy modification).
          *
          * @private
          * @method _getStoredEndpoints
@@ -718,6 +741,9 @@ watcher = stampit().state({
             var endpoint = this._registry[id];
             if (endpoint) {
                 var now = moment().utc();
+                if (undetermined !== endpoint.status) {
+                    this._storeStatusToHistory(endpoint, undetermined);
+                }
                 endpoint.active = active;
                 endpoint.status = undetermined;
                 endpoint.timestamp = now;
@@ -775,6 +801,45 @@ watcher = stampit().state({
         },
 
         /**
+         * Returns the endpoint/s history status.
+         *
+         * @method getHistory
+         * @param {Object} qOpts the query options.
+         * @param {Function} callback the callback(err, recs) of the async process.
+         * @example __Query options details__.
+         * ```
+         * //Specify the "endpointId" to return the history of the specific endpoint,
+         * //otherwise the history of all endpoints is returned
+         * //Specify both 'from' and 'to' properties to return the history status for
+         * //the specific time period, otherwise the complete history is returned.
+         *
+         * {
+         *     endpointId:  <endpoint id>
+         *     from:        <from time>     //defined in mills
+         *     to:          <to time>       //defined in mills
+         * }
+         * ```
+         */
+        getHistory: function getHistory(qOpts, callback) {
+            var query = {};
+            var endpointId = qOpts.endpointId;
+            var from = qOpts.from;
+            var to = qOpts.to;
+            if (from && to) {
+                query.timestamp = {$gte: parseFloat(from), $lte: parseFloat(to)};
+            }
+            if (endpointId) {
+                query.endpointId = endpointId;
+            }
+            this.db.history.find({$query: query, $orderby: {timestamp: 1}}, function (err, recs) {
+                if (err) {
+                    logger.error('Unable to get history, err: ' + err + ', query: ' + query);
+                }
+                callback(err, recs);
+            });
+        },
+
+        /**
          * Setup the Watcher.
          *
          * @private
@@ -806,6 +871,10 @@ watcher = stampit().state({
                         app.post('/endpoints/:id/notify', routes.endpointEnableNotification(_self));
                         app.delete('/endpoints/:id/notify', routes.endpointDisableNotification(_self));
                         app.get('/resolution-strategies', routes.resolutionStrategies(_self));
+                        app.get('/history/endpoints', routes.history(_self));
+                        app.get('/history/endpoints/:id', routes.history(_self));
+                        app.get('/history/endpoints/:from/:to', routes.history(_self));
+                        app.get('/history/endpoints/:id/:from/:to', routes.history(_self));
                         app.get('/console', function (req, res) {
                             res.render('console');
                         });
@@ -861,7 +930,7 @@ watcher = stampit().state({
                 host + '/' + port + '/' + interval + ')');
                 setTimeout(_.bind(_self._pollEndpoints, _self), 0).unref();
             });
-
+            _self.db = dbConnFactory.create(_self.options.dbConnectionURL);
             return _self;
         }),
 
@@ -871,9 +940,42 @@ watcher = stampit().state({
          * @method stop
          */
         stop: _.once(function stop() {
-            logger.debug('Shutting down Watcher app...');
-            this._stopped = true;
-            this.server.stop();
+            logger.info('Shutting down Watcher app...');
+            var _self = this;
+            _self._stopped = true;
+            async.parallel([
+                function (callback) {
+                    //insert new entries with 'undetermined'
+                    var entries = [];
+                    var now = moment.utc();
+                    _.each(_self.getEndpoints(), function (endpoint) {
+                        if (undetermined !== endpoint.status) {
+                            entries.push({
+                                endpointId: endpoint.id,
+                                timestamp: now.valueOf(),
+                                statusTransition: {
+                                    from: endpoint.status,
+                                    to: undetermined
+                                }
+                            });
+                        }
+                    });
+                    _self.db.history.insert(entries, function (err, docs) {
+                        callback(err);
+                        _self.db.close();
+                    });
+                },
+                asyncTimeout(function (callback) {
+                    _self.server.stop();
+                }, 2000, 'Give 2 secs to shutdown http server')
+            ], function (err, results) {
+                if (err) {
+                    logger.info('Shutdown performed with errors: ' + err);
+                } else {
+                    logger.info('Shutdown successfully performed');
+                }
+                process.exit();
+            });
             //process.kill(process.pid, 'SIGHUP');
         }),
 
@@ -902,6 +1004,7 @@ watcherFactory = {
      * @static
      * @method create
      * @param {Object} [options] the Watcher configuration.
+     * @example __Options configuration details__.
      * Properties in _[]_ are optional. when not set, the default values are used - those in parentheses.
      * * Embedded http server configuration and service communication interval
      *  * [__host__] ('localhost'), the http server host name.
@@ -927,6 +1030,10 @@ watcherFactory = {
      *  }
      *  ```
      *  * [__exportDir__] (_<`project directory`>/storage/_), the storage directory of dynamically created endpoints.
+     *  * [__dbConnectionURL__] (_'mongodb://localhost:27017/'_), the connection url of history persistent storage
+     *  (currently, the mongoDB is used)
+     *
+     *
      * * Endpoint
      *  * __id__, the unique endpoint id.
      *  * __type__, the endpoint type.
@@ -954,6 +1061,7 @@ watcherFactory = {
             resolutionStrategies: [],
             nfOpts: {},
             exportDir: __dirname + '/../storage',
+            dbConnectionURL: 'mongodb://localhost:27017/',
             endpoints: []
         };
         options = options || {};