+++ test/test-watcher.js
@@ -4,9 +4,24 @@
 'use strict';
 var http = require('http');
 var s = require('underscore.string');
+var _ = require('underscore');
 var assert = require('chai').assert;
+var proxyquire = require('proxyquire');
 
-var app = require('../src/watcher');
+var stubs = {
+    './database': {
+        create: function create() {
+            return {
+                history: {
+                    insert: _.noop
+                }
+            };
+        },
+        '@global': true
+    }
+};
+
+var app = proxyquire('../src/watcher', stubs);
 
 describe('watcher', function () {
     describe('#test start watcher', function () {